+++ vaadin-context-menu.d.ts
@@ -1,2 +1,2 @@
 export * from './src/vaadin-context-menu.js';
-export * from './@types/interfaces';
+export * from './@src/interfaces';
